Fix build errors: ConnectionStatus import, Icons.road, RadarEntry type, flutter_local_notifications
- dashboard_screen.dart: add missing websocket_service.dart import for ConnectionStatus
- dashboard_screen.dart: Icons.road → Icons.route (road icon doesn't exist)
- report_screen.dart: cast RadarEntry value to double (num not accepted)
- gradle.properties: disable Kotlin incremental compilation (cross-drive bug on Windows)
- pubspec.yaml: upgrade flutter_local_notifications ^16.3.0 → ^18.0.0 (fixes bigLargeIcon ambiguity on Android API 31+)

+++ mobile/lib/features/reports/report_screen.dart
@@ -122,7 +122,7 @@ class _ScoreRadar extends StatelessWidget {
                 dataSets: [
                   RadarDataSet(
                     dataEntries: scores
-                        .map((s) => RadarEntry(value: s))
+                        .map((s) => RadarEntry(value: s.toDouble()))
                         .toList(),
                     fillColor: AppTheme.neonCyan.withOpacity(0.15),
                     borderColor: AppTheme.neonCyan,
